adjust hook data validations (#33)
Addresses 5.1: "Hook Data Format Not Enforced"

Audit item discusses three points:
1. The magic bytes "cctp-forward" are not checked
2. Setting the magic bytes to 0 does not opt out of forwarding.
3. The length of the Circle Hook Data is not checked; instead, the
function always reads 20 bytes for the forwardRecipient address.

(1) Setting the magic bytes should not be required by this contract.
However, it will be used by Iris to identify which messages to forward
via this contract. I have updated the comment description to reflect
this
(2) Setting magic bytes to 0 will opt out of forwarding _by Circle_. I
have clarified that in the comment. It is not possible to opt out any
possibility of forwarding, because we want users to be able to forward
by themselves if Circle does not forward.
(3) The "length" part of the hook message is ignored by this contract,
but the actual length of hook bytes is validated. Specifically, we were
validating that it was 52 bytes. I have updated this validation to check
that the hook is at least 52 bytes, in order to support forwarding hooks
which conform to our spec and also add arbitary data. This data is not
returned or used by our contracts, but it may enable some additional
composable flows (e.g. another contract validating that part of the hook
before passing it to our forwarder), and is easy to support, so I think
it is worthwhile.